Move GelfMock to a separate file
