+++ spec/spec_helper.rb
@@ -15,6 +15,7 @@ SimpleCov.start
 
 require 'httplog'
 
+require 'loggers/gelf_mock'
 require 'adapters/http_base_adapter'
 Dir[File.dirname(__FILE__) + '/adapters/*.rb'].each { |f| require f }
 Dir['./spec/support/**/*.rb'].each { |f| require f }
@@ -38,21 +39,5 @@ RSpec.configure do |config|
 
   def log
     @log.string
-  end
-end
-
-class GelfMock < Logger
-  def log(severity, message = nil, progname = nil)
-    message ||= {}
-    if message.is_a?(Hash)
-      message[:short_message] = message[:short_message].to_s
-
-      message = message.each_with_object({}) do |(key, value), obj|
-        key_s = key.to_s
-
-        obj[key_s] = value
-      end.to_json
-    end
-    super(severity, message, progname)
   end
 end
